requirements.txt
mysql-connector-python
python-dotenv
llama-index-core
llama-parse
sec-edgar-downloader
sec-downloader
streamlit
openai
tiktoken
requests
beautifulsoup4
pandas
tavily-python
aiohttp
aiohttp-cors
uvicorn
yfinance
